[Silver III] Title: 주유소, Time: 580 ms, Memory: 43768 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver III/13305. 주유소/주유소.java
@@ -24,9 +24,9 @@ class Main {
         writer.close();
     }
 
-    private static int calculateMinimumOilCost(int[] distance, int[] costsByCity) {
-        int minimumCost = costsByCity[0];
-        int total = minimumCost * distance[0];
+    private static long calculateMinimumOilCost(int[] distance, int[] costsByCity) {
+        long minimumCost = costsByCity[0];
+        long total = minimumCost * distance[0];
         for (int i = 1; i < distance.length; i++) {
             if (minimumCost > costsByCity[i]) minimumCost = costsByCity[i];
             total += minimumCost * distance[i];
